chore: organized code
