+++ src/TailwindVariants.NET/Models/CompiledVariant.cs
@@ -5,13 +5,9 @@ using static TailwindVariants.NET.TvHelpers;
 
 namespace TailwindVariants.NET;
 
-public interface IApplicableVariant
-{
-	void Apply(object owner, Action<string, string> aggregator);
-}
-
-public interface ICompiledCompoundVariant : IApplicableVariant;
-
+/// <summary>
+/// Represents a compiled single variant, which applies based on a single variant condition.
+/// </summary>
 public interface ICompiledVariant : IApplicableVariant;
 
 internal record struct CompiledVariant<TOwner, TSlots>(Expression<VariantAccessor<TOwner>> Expr, IVariant<TSlots> Entry, VariantAccessor<TOwner> Accessor) : ICompiledVariant
@@ -36,37 +32,6 @@ internal record struct CompiledVariant<TOwner, TSlots>(Expression<VariantAccesso
 		catch (Exception ex)
 		{
 			Debug.WriteLine($"Variant evaluation failed for '{Expr}': {ex.Message}");
-		}
-	}
-}
-
-internal record struct CompiledCompoundVariant<TOwner, TSlots>(Predicate<TOwner> Predicate, SlotCollection<TSlots> Slots) : ICompiledCompoundVariant
-	where TSlots : ISlots, new()
-	where TOwner : ISlotted<TSlots>
-{
-	public readonly void Apply(object owner, Action<string, string> aggregator)
-	{
-		try
-		{
-			if (!Predicate((TOwner)owner))
-			{
-				return;
-			}
-
-			foreach (var (slot, value) in Slots)
-			{
-				if (slot is null)
-				{
-					continue;
-				}
-
-				aggregator(GetSlot(slot), value.ToString());
-			}
-		}
-		catch (Exception ex)
-		{
-			// keep robust but log for debugging
-			Debug.WriteLine($"Compound variant predicate or processing failed: {ex.Message}");
 		}
 	}
 }
